+++ pedersen/src/lib.rs
@@ -1,12 +1,12 @@
 #![forbid(unsafe_code)]
-pub mod ec_point_add_protocol;
+pub mod cdls_point_add_protocol;
+pub mod cdls_scalar_mul_protocol;
 pub mod equality_protocol;
-pub mod fs_scalar_mul_protocol;
+pub mod fs_cdls_scalar_mul_protocol;
 pub mod fs_zk_attest_scalar_mul_protocol;
 pub mod mul_protocol;
 pub mod opening_protocol;
 pub mod pedersen_config;
-pub mod scalar_mul_protocol;
 pub mod transcript;
 pub mod zk_attest_point_add_protocol;
 pub mod zk_attest_scalar_mul_protocol;
